Demonstrate a multi-module setup in Rollup
Introduces an additional module that uses Vue and WVUI. These are
both external dependencies provided by MediaWiki, so they are
loaded with "require" instead of "import".

rollup-plugin-vue is necessary to transform Vue files in this setup.

The rollup.config.js file has been modified to export an array of
config objects in order to support this multiple entries/multiple outputs
scenario.