Integrate dev mode with ResourceLoader through a callback
Also add the less npm package so that less works.

This still has the issue that Vue is not externalized in dev mode (so
Vite's version of Vue is used instead of MW's, so the i18n plugin isn't
loaded), and that Vite's less compiler is used rather than MW's (so MW's
import path isn't honored, and custom LESS vars don't work).